Move ListToolbar to own component "Toolbar" under PortfolioWebParts
