+++ SharePointFramework/PortfolioWebParts/src/components/List/types.ts
@@ -2,7 +2,7 @@ import { IColumn, IScrollablePaneProps, IShimmeredDetailsListProps, Target } fro
 import { SearchBoxProps } from '@fluentui/react-search-preview'
 import { WebPartContext } from '@microsoft/sp-webpart-base'
 import { IFilterPanelProps } from 'pp365-shared-library'
-import { ListMenuItem } from './ListToolbar'
+import { ListMenuItem } from '../Toolbar'
 
 export type OnColumnContextMenu = {
   column: any
