+++ SharePointFramework/PortfolioWebParts/src/components/Toolbar/createIcon.tsx
