+++ SharePointFramework/PortfolioWebParts/src/components/Toolbar/createStyle.ts
