+++ SharePointFramework/PortfolioWebParts/src/components/Toolbar/index.tsx
@@ -1,16 +1,17 @@
-import { Toolbar } from '@fluentui/react-components'
+import { Toolbar as ReactToolbar } from '@fluentui/react-components'
 import strings from 'PortfolioWebPartsStrings'
 import { FilterPanel } from 'pp365-shared-library'
 import React, { FC } from 'react'
-import { useListContext } from '../context'
+import { useToolbarContext } from './context'
 import { renderToolbarItem } from './renderToolbarItem'
 import styles from './ListToolbar.module.scss'
+import { IToolbarProps } from './types'
 
-export const ListToolbar: FC = () => {
-  const context = useListContext()
+export const Toolbar: FC<IToolbarProps> = () => {
+  const context = useToolbarContext()
   return (
     <div>
-      <Toolbar className={styles.toolbar}>{context.props.menuItems.map(renderToolbarItem)}</Toolbar>
+      <ReactToolbar className={styles.toolbar}>{context.props.menuItems.map(renderToolbarItem)}</ReactToolbar>
       <FilterPanel
         {...context.props.filterPanelProps}
         headerText={strings.FiltersString}
