+++ SharePointFramework/PortfolioWebParts/src/components/Toolbar/types.ts
@@ -1,5 +1,19 @@
 import { FluentIcon } from '@fluentui/react-icons/lib/utils/createFluentIcon'
+import { IFilterPanelProps } from 'pp365-shared-library'
 import { CSSProperties, MouseEventHandler } from 'react'
+
+export interface IToolbarProps {
+  /**
+   * Menu items to render in the Toolbar.
+   */
+  menuItems?: ListMenuItem[]
+
+  /**
+   * Filter panel props.
+   */
+  filterPanelProps?: IFilterPanelProps
+
+}
 
 /**
  * Represents a menu item in the `ListMenuItem` component. Supports
